TraktCommentsAdapter: round avatar images.
